refactor: use enumerate devices hook
